Move multer handler to routing module
